refactor: separate method body binder
